+++ types/block_test.go
@@ -464,11 +464,11 @@ func TestBlockMaxDataBytes(t *testing.T) {
 	}{
 		0: {-10, 1, 0, true, 0},
 		1: {10, 1, 0, true, 0},
-		2: {841, 1, 0, true, 0},
-		3: {842, 1, 0, false, 0},
-		4: {843, 1, 0, false, 1},
-		5: {954, 2, 0, false, 1},
-		6: {1053, 2, 100, false, 0},
+		2: {4088, 1, 0, true, 0},
+		3: {4089, 1, 0, false, 0},
+		4: {4090, 1, 0, false, 1},
+		5: {7448, 2, 0, false, 1},
+		6: {7547, 2, 100, false, 0},
 	}
 
 	for i, tc := range testCases {
@@ -495,9 +495,9 @@ func TestBlockMaxDataBytesNoEvidence(t *testing.T) {
 	}{
 		0: {-10, 1, true, 0},
 		1: {10, 1, true, 0},
-		2: {841, 1, true, 0},
-		3: {842, 1, false, 0},
-		4: {843, 1, false, 1},
+		2: {4088, 1, true, 0},
+		3: {4089, 1, false, 0},
+		4: {4090, 1, false, 1},
 	}
 
 	for i, tc := range testCases {