+++ types/signable.go
@@ -2,6 +2,7 @@ package types
 
 import (
 	"github.com/cometbft/cometbft/crypto/ed25519"
+	"github.com/cometbft/cometbft/crypto/mldsa65"
 	cmtmath "github.com/cometbft/cometbft/libs/math"
 )
 
@@ -9,7 +10,7 @@ var (
 	// MaxSignatureSize is a maximum allowed signature size for the Proposal
 	// and Vote.
 	// XXX: secp256k1 does not have Size nor MaxSize defined.
-	MaxSignatureSize = cmtmath.MaxInt(ed25519.SignatureSize, 64)
+	MaxSignatureSize = cmtmath.MaxInt(ed25519.SignatureSize, mldsa65.SignatureSize)
 )
 
 // Signable is an interface for all signable things.